# [SPARK-42551][SQL] Support more subexpression elimination cases