--- a/sql/core/src/main/scala/org/apache/spark/sql/execution/joins/JoinCodegenSupport.scala
+++ b/sql/core/src/main/scala/org/apache/spark/sql/execution/joins/JoinCodegenSupport.scala
@@ -49,11 +49,14 @@ trait JoinCodegenSupport extends CodegenSupport with BaseJoinExec {
 
       // filter the output via condition
       ctx.currentVars = streamVars ++ buildVars
-      val ev =
-        BindReferences.bindReference(expr, streamPlan.output ++ buildPlan.output).genCode(ctx)
+      val bondExpr = BindReferences.bindReference(expr, streamPlan.output ++ buildPlan.output)
+      initBlock += ctx.subexpressionElimination(bondExpr)
+      val ev = bondExpr.genCode(ctx)
+
       val skipRow = s"${ev.isNull} || !${ev.value}"
       s"""
          |$eval
+         |$initBlock
          |${ev.code}
          |if (!($skipRow))
        """.stripMargin