@wankunde 2023-03-01
Hi, @cloud-fan, unlike ProjectExec and HashAggregateExec, FilterExec should handle subexpression elimination in a special way. So if any of the filter expressions return false, we will be able to skip evaluating the remaining common expressions.

Could we support CSE for the other operations in the following PR?