@wankunde 2023-03-03
Hi, @cloud-fan @viirya 
I updated the PR and support subexpression elimination in FilterExec and JoinExec.
We can support some other operator subexpression elimination in two steps.
Do you have any good advice?