@wankunde 2023-04-14
Use https://github.com/apache/spark/pull/40713 instead of this PR, close it.